requirements.txt
google-cloud-texttospeech
google-cloud-speech
pygame
